break out ApplicationSerializer
